Refactor to separte friendship and invitations
